+++ src/internet/model/nsc-tcp-l4-protocol.h
@@ -48,9 +48,9 @@ class NscTcpL4Protocol : public IpL4Protocol {
 public:
 
 #ifndef _MSC_VER
-        static const uint16_t PROT_NUMBER;//!< protocol number (0x6)
+        static const uint8_t PROT_NUMBER;//!< protocol number (0x6)
 #else
-        constexpr static const uint16_t PROT_NUMBER = 0x6;
+        constexpr static const uint8_t PROT_NUMBER = 0x6;
 #endif
   /**
    * \brief Get the type ID.
